+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/magical/StormStaff.java
@@ -9,6 +9,7 @@ import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.Sound;
+import org.bukkit.entity.LightningStrike;
 import org.bukkit.entity.Player;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
 import org.bukkit.inventory.ItemStack;
@@ -21,9 +22,17 @@ import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Lists.SlimefunItems;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.handlers.ItemUseHandler;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
+/**
+ * This {@link SlimefunItem} casts a {@link LightningStrike} where you are pointing.
+ * Unlike the other Staves, it has a limited amount of uses.
+ * 
+ * @author Linox
+ *
+ */
 public class StormStaff extends SimpleSlimefunItem<ItemUseHandler> {
 
     public static final int MAX_USES = 8;