+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/ItemState.java
@@ -1,24 +1,30 @@
 package me.mrCookieSlime.Slimefun.Objects.SlimefunItem;
 
+import io.github.thebusybiscuit.slimefun4.implementation.items.VanillaItem;
+
 /**
  * Defines whether a SlimefunItem is enabled, disabled or fall-back to its vanilla behavior.
  *
  * @author Poslovitch
+ * 
+ * @see SlimefunItem
+ * 
  */
 public enum ItemState {
 
     /**
-     * This SlimefunItem is enabled.
+     * This {@link SlimefunItem} is enabled.
      */
     ENABLED,
 
     /**
-     * This SlimefunItem is disabled and is not a {@link VanillaItem}.
+     * This {@link SlimefunItem} is disabled and is not a {@link VanillaItem}.
      */
     DISABLED,
 
     /**
-     * This SlimefunItem is fall-back to its vanilla behavior, because it is disabled and is a {@link VanillaItem}.
+     * This {@link SlimefunItem} has fallen back to its vanilla behavior, because it is disabled and an instance of
+     * {@link VanillaItem}.
      */
     VANILLA
 }